# echango-api-main/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/v1/offices/{id}:
    get:
      operationId: OfficesController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
      security:
        - bearer: []
      tags: *ref_2
    patch:
      operationId: OfficesController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfficeDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
      security:
        - bearer: []
      tags: *ref_2
    delete:
      operationId: OfficesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_2
...
  /api/v1/files/upload:
    post:
      operationId: FilesLocalController_uploadFile_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: >-
            File upload - returns single object for one file, array for multiple
            files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponseDto'
      security:
        - bearer: []
      tags: &ref_6
        - Files
  /api/v1/files/upload-multiple:
    post:
      operationId: FilesLocalController_uploadFiles_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileResponseDto'
      security:
        - bearer: []
      tags: *ref_6
...
  /api/v1/office-rates:
    post:
      operationId: OfficeRatesController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfficeRateDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeRate'
      security:
        - bearer: []
      tags: &ref_9
        - Officerates
    get:
      operationId: OfficeRatesController_getOfficeRatesByOfficeId_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_9
  /api/v1/office-rates/rate-management:
    post:
      operationId: OfficeRatesController_createRateManagement_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfficeRateDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeRate'
      security:
        - bearer: []
      tags: *ref_9
  /api/v1/office-rates/{id}:
    get:
      operationId: OfficeRatesController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeRate'
      tags: *ref_9
    patch:
      operationId: OfficeRatesController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfficeRateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeRate'
      security:
        - bearer: []
      tags: *ref_9
    delete:
      operationId: OfficeRatesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_9
  /api/v1/office-rates/rate-management/{id}:
    patch:
      operationId: OfficeRatesController_updateRateManagement_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfficeRateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeRate'
      security:
        - bearer: []
      tags: *ref_9
  /api/v1/office-rates/bulk-update:
    post:
      description: >-
        Updates or creates currency rates for multiple offices identified by
        their slugs. If a rate exists for a currency, it will be updated. If
        not, a new rate will be created.
      operationId: OfficeRatesController_bulkUpdateOfficeRates_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkUpdateOfficeRatesDto'
      responses:
        '201':
          description: Bulk update completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  results:
                    type: object
                    properties:
                      updated:
                        type: number
                      created:
                        type: number
                      errors:
                        type: number
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        officeSlug:
                          type: string
                        currency:
                          type: string
                        action:
                          type: string
                          enum:
                            - updated
                            - created
                            - error
                        message:
                          type: string
      summary: Bulk update office rates for multiple offices
      tags: *ref_9
...
  /api/v1/notification-preferences/me:
    get:
      description: Get the current user notification preferences
      operationId: NotificationPreferencesController_getMyPreferences_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
      security: &ref_21
        - bearer: []
      summary: Get my notification preferences
      tags: &ref_22
        - Notification Preferences
    patch:
      description: Update the current user notification preferences
      operationId: NotificationPreferencesController_updateMyPreferences_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationPreferenceDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
      security: *ref_21
      summary: Update my notification preferences
      tags: *ref_22
...
info:
  title: ExchangGo24 API
  description: ' API docs'
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    FileType:
      type: object
      properties:
        id:
          type: string
          example: cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae
        path:
          type: string
          example: https://example.com/path/to/file.jpg
      required:
        - id
        - path
    Role:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: admin
      required:
        - id
        - name
    Status:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: active
      required:
        - id
        - name
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
          example: john.doe@example.com
        provider:
          type: string
          example: email
        socialId:
          type: string
          example: '1234567890'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        photo:
          $ref: '#/components/schemas/FileType'
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - provider
        - socialId
        - firstName
        - lastName
        - photo
        - role
        - status
        - createdAt
        - updatedAt
        - deletedAt
    WorkingHour:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the working hour record
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dayOfWeek:
          type: string
          description: Day of the week
          example: Monday
          enum:
            - Monday
            - Tuesday
            - Wednesday
            - Thursday
            - Friday
            - Saturday
            - Sunday
        isActive:
          type: boolean
          description: Indicates if the day is active for working hours
          example: true
        fromTime:
          type: string
          description: Start time of working hours
          example: '09:00'
          nullable: true
        toTime:
          type: string
          description: End time of working hours
          example: '17:00'
          nullable: true
        hasBreak:
          type: boolean
          description: Indicates if there is a break during this day
          example: false
        breakFromTime:
          type: string
          description: Start time of break
          example: '12:00'
          nullable: true
        breakToTime:
          type: string
          description: End time of break
          example: '13:00'
          nullable: true
        officeId:
          type: string
          description: ID of the associated office
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: Creation timestamp
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          format: date-time
          type: string
          description: Last update timestamp
          example: '2023-01-02T12:00:00Z'
      required:
        - id
        - dayOfWeek
        - isActive
        - fromTime
        - toTime
        - hasBreak
        - breakFromTime
        - breakToTime
        - officeId
        - createdAt
        - updatedAt
    Office:
      type: object
      properties:
        id:
          type: string
        owner:
          $ref: '#/components/schemas/User'
        officeName:
          type: string
        location:
          type: object
        registrationNumber:
          type: string
        currencyExchangeLicenseNumber:
          type: string
        address:
          type: string
        city:
          type: object
        state:
          type: object
        country:
          type: object
        primaryPhoneNumber:
          type: string
        secondaryPhoneNumber:
          type: object
        thirdPhoneNumber:
          type: object
        whatsappNumber:
          type: object
        logo:
          type: object
        mainImage:
          type: object
        slug:
          type: object
        isActive:
          type: boolean
        isVerified:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          type: object
        distanceInKm:
          type: number
        rates:
          type: array
          items:
            type: string
        equivalentValue:
          type: number
        images:
          type: object
        workingHours:
          type: array
          items:
            type: string
        email:
          type: object
        todayWorkingHours:
          $ref: '#/components/schemas/WorkingHour'
      required:
        - id
        - owner
        - officeName
        - location
        - registrationNumber
        - currencyExchangeLicenseNumber
        - address
        - city
        - state
        - country
        - primaryPhoneNumber
        - secondaryPhoneNumber
        - thirdPhoneNumber
        - whatsappNumber
        - logo
        - mainImage
        - slug
        - isActive
        - isVerified
        - isFeatured
        - createdAt
        - updatedAt
        - deletedAt
        - distanceInKm
        - rates
        - equivalentValue
        - images
        - workingHours
        - email
        - todayWorkingHours
...
    FileDto:
      type: object
      properties:
        id:
          type: string
      required:
        - id
...
    UpdateOfficeDto:
      type: object
      properties:
        officeName:
          type: string
          example: Office Name
        registrationNumber:
          type: string
          example: AAB-1234567890
        currencyExchangeLicenseNumber:
          type: string
          example: AAB-1234567890
        address:
          type: string
          example: 123 Main St, Rabat, Morocco
        city:
          type: string
          example: Rabat
        state:
          type: string
          example: Rabat-Salé
        primaryPhoneNumber:
          type: string
          example: +212 661 23 45 67
        secondaryPhoneNumber:
          type: string
          example: +212 661 23 45 67
        thirdPhoneNumber:
          type: string
          example: +212 661 23 45 67
        whatsappNumber:
          type: string
          example: +212 661 23 45 67
        logo:
          example: logo.png
          allOf:
            - $ref: '#/components/schemas/FileDto'
        mainImage:
          example: main-image.jpg
          allOf:
            - $ref: '#/components/schemas/FileDto'
        location:
          type: object
          example: '{ type: "Point", coordinates: [longitude, latitude] }'
...
    Currency:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        namePlural:
          type: string
        symbol:
          type: string
        symbolNative:
          type: string
        decimalDigits:
          type: number
        rounding:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        flag:
          type: string
      required:
        - id
        - code
        - name
        - namePlural
        - symbol
        - symbolNative
        - decimalDigits
        - rounding
        - createdAt
        - updatedAt
        - flag
...
    FileResponseDto:
      type: object
      properties:
        file:
          $ref: '#/components/schemas/FileType'
      required:
        - file
...
    CreateOfficeRateDto:
      type: object
      properties:
        targetCurrency:
          type: string
          example: USD
        buyRate:
          type: number
          example: 10
        sellRate:
          type: number
          example: 10
        isActive:
          type: boolean
          example: true
      required:
        - targetCurrency
        - buyRate
        - sellRate
        - isActive
    OfficeRate:
      type: object
      properties:
        id:
          type: string
        office:
          $ref: '#/components/schemas/Office'
        buyRate:
          type: number
        sellRate:
          type: number
        baseCurrency:
          $ref: '#/components/schemas/Currency'
        targetCurrency:
          $ref: '#/components/schemas/Currency'
        isActive:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          type: object
        buyRateEquivalent:
          type: number
        sellRateEquivalent:
          type: number
        equivalentValue:
          type: number
      required:
        - id
        - office
        - buyRate
        - sellRate
        - baseCurrency
        - targetCurrency
        - isActive
        - createdAt
        - updatedAt
        - deletedAt
        - buyRateEquivalent
        - sellRateEquivalent
        - equivalentValue
    UpdateOfficeRateDto:
      type: object
      properties: {}
    CurrencyRateDto:
      type: object
      properties:
        currency:
          type: string
          example: EUR
          description: Currency code (3 letters)
        buy:
          type: number
          example: 10.3
          description: Buy rate for the currency
        sell:
          type: number
          example: 10.65
          description: Sell rate for the currency
      required:
        - currency
        - buy
        - sell
    BulkUpdateOfficeRatesDto:
      type: object
      properties:
        rates:
          description: Array of currency rates to update
          example:
            - currency: EUR
              buy: 10.3
              sell: 10.65
            - currency: USD
              buy: 8.85
              sell: 9.25
          type: array
          items:
            $ref: '#/components/schemas/CurrencyRateDto'
        officeSlugs:
          description: Array of office slugs to update
          example:
            - change-goulmima-exchange-7j7
            - el-jadida-cash
            - saraf-al-inara
            - change-boulevard-roudani-change-maarif-casablanca
          type: array
          items:
            type: string
      required:
        - rates
        - officeSlugs
...
    NotificationPreference:
      type: object
      properties:
        id:
          type: string
        rateUpdateReminderWhatsApp:
          type: boolean
          description: Enable WhatsApp notifications for rate update reminders
        rateUpdateReminderEmail:
          type: boolean
          description: Enable email notifications for rate update reminders
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - rateUpdateReminderWhatsApp
        - rateUpdateReminderEmail
        - createdAt
        - updatedAt
    UpdateNotificationPreferenceDto:
      type: object
      properties:
        rateUpdateReminderWhatsApp:
          type: boolean
          description: Enable WhatsApp notifications for rate update reminders
          example: true
        rateUpdateReminderEmail:
          type: boolean
          description: Enable email notifications for rate update reminders
          example: true
...